Fixed the bash prompt length and added $? parens
The bash prompt now uses \[ and \] to tell bash not to count terminal
control characters when trying to figure out how wide the prompt is,
correcting a terminal formatting error; and if the previous command
produced an error, then the $? exit status is now shown in parens.
Also, the prompt is no longer even installed if TERM is not set, to
prevent "tput" errors from being printed when I am ssh'ing a file!