HTML page I can bookmark and use to clean up URLs
People sometimes email URLs to me.  Given the length of typical URLs,
they often take up several lines in my terminal, and Emacs adds a
backslash and newline at the end of each line which make the URL
invalid.  This tiny local web page, which I can bookmark in Google
Chrome as a file:// link, lets me paste in such a URL and fixes it.