Prevent window from "losing" borders upon Win+Tab
Before, pressing the Win key with Tab to vertically toggle a window
would fail to give the window its borders back if an F11 (for a browser)
or Win+F (for any window) had fullscreened the window by maximizing it
without borders. Now, Win+Tab always brings the window back to normal
size with its borders instead of without.