No longer update homedir's pip directly from PyPI
Installing a newer pip version during the install process was resulting
in an error "pkg_resources.DistributionNotFound: pip==1.3.1" because
this script would install pip 1.4 in place of pip 1.3.1.  So I no longer
automatically upgrade pip as part of the setup script.

Also, removed bpython, because it was taking too long to install and I
never really use it anyway.