Make it easy to run windowed Emacs from the shell
For years I have simply used my window manager shortcut [Windows]+[E] to
run an Emacs, but now that I am using the Jedi completion system I need
a way to run an Emacs that has the same environment variables set as my
current shell, so that it can see the $VIRTUAL_ENV setting for the
project I am currently working on.  So I am bringing back my old "e"
alias, and giving it a fallback of in-terminal Emacs if the windowed
version does not seem to be possible.