Rewrite ,simplehttpserver in Python to target 8002
It was annoying to always run SimpleHTTPServer on a random port because
then my browser history gets spammed with a different localhost:<n> each
time.  So from now on I want SimpleHTTPServer to simply grab the first
available port starting at 8002 so that a small number of ports will
keep getting re-used.