Restyle windows: no borders, make titlebar black
Actually there is a very slight border of 1px so that I can see the
edges of windows.  The window titlebar is not only black (which will
help window borders and titles be more visible when I am hooked up to
projectors in front of an audience), but I have moved the titles to the
upper-right-hand corner to put them in a predictable position that is
nevertheless out of the way visually.

Since there is now nothing to grab to resize a window at its edge, I
have made middle-click on the titlebar enact a unilateral resize action.