Switch to natural www.rhodesmill.org hostname
I suppose that I was using the IP address for robustness in the face of
DNS failure?  But the result is that the move of the server to new
hardware and IP meant that my alias for it stopped working.  So let's
try the hostname for a while and see whether it works reliably or not.