Install fewer things in Anaconda environments
By only installing a few things by default, it is now possible for me to
create Python 3 environments; before, the sheer weight of dependencies
was forcing all Anaconda environments to be Python 2 even if I put
"python=3" on the command line.

Also, this makes a "conda" command available when an Anaconda
environment has been activated.