Comment out Docker install after return to 32-bit
My Latitude D630 is simply not up to the task of running a 64-bit OS.
Just a half-dozen busy Chrome tabs open, and the system starts swapping
and falls to its knees.  So I am returning to 32-bit for the main system
install so that Chrome keeps working, and will do my Docker experiments
inside a VirtualBox-hosted 64-bit Ubuntu machine that does nothing
challenging like browsing the web.