Change where conda environments are stored
The conda tool now wants environment names to have no slashes inside,
and defaults to wanting environments to live in ~/.anaconda/envs.  I am
making an attempt to go along with it, although I fear that I might soon
tire of having to go one place to dispose of virtualenvs (~/.v) and
another to dispose of environments and reclaim their disk space.  We
will see.