Remove the need for a ".gitmessage.txt" file
Back when I used to make fairly scripted commits to a particular
project, it helped to force myself to have a pre-written git message
template that I could fill out each time.  But those days are now past,
and it is annoying to get the error:

    fatal: could not read '/home/brandon/.gitmessage.txt': No such file or directory

each time I try to commit with `git` on a new account or machine.