Move browser choice to `,up` script
I run `,setup-ubuntu` rarely, as it is fairly heavyweight and requires
my password, so I am moving simpler settings and cleanup that has to do
with my account, rather than the whole system, into a little `,up`
script that I can run all the time to re-assert settings that might go
awry.  And my choice of browser clearly belongs there, as it can be
damaged later by installers and so forth, but should be repairable
without typing my password and waiting for my whole Ubuntu setup script
to run.

Also, while moving the settings, I am fixing them: it now looks like
`chromium-browser.desktop` is the name that I need to be using, not just
the bare `chromium.desktop`.  Everything was opening in Firefox in the
meantime.