Blank screen go blank 1 minute before lock
The little notify message in the corner has not been enough.  I often
don't see it in time, and the screen locks while I am in the middle of a
Hangout, or watching a video.  So, something more obvious, and that
gives me a longer window before the screen locks: 60 seconds before the
screen locks, the screen will turn off.  It will be harder to miss that.