Have `git lp` auto-search for `^CommitDate:`
The biggest problem with `git log -p` is it is so verbose, and it can be
hard to find the start of the next diff amidst all the code.  So now
when I run `git lp`, I will have `less` go ahead and start by searching
for a pattern that only ever occurs in the header of a diff, so I can
then use `n` and `N` to find the next and previous diff.  Setting the
`LESSHISTFILE` prevents this search term from overwriting my previous
hand-typed search term.