Switch to xterm for popup status screen
Before, when I wanted to know the date or time, I was hiding all windows
to reveal the background, which I have been keeping updated with that
information every 60 seconds.  But hiding and revealing all windows is
disruptive, so this tries something else: having a popup xterm appear on
the fly when I want the information.