Have xterm grab focus when bell is rung
This allows tools needing to prompt me to ring the terminal bell to
demand focus, instead of having me assume that a long task is still
running and having me only discover much later that the xterm is still
sitting at a prompt.  Also, let's go ahead and turn on the visual bell
so that I can see whether I like being able to see when it is rung.