Switch to a script to launch editors, terminals
I used to hard-code the windows I wanted open when working with each
specific project, but that doesn’t scale once there are enough projects.
Instead, this new “,p” script introspects the current project and tries
to launch all the right things (a list that will grow as I live with it
over the next week or two).