Bring back ability to shortcut-key between windows
Moving with the trackpoint is almost painful on my Lenovo Thinkpad X260,
even with the sensitivity turned all the way up; and given that the
trackpoint’s response is asymmetric in different directions, I can only
conclude that the problem is one of manufacturing versus earlier
Thinkpads.  Especially while I still use it as a laptop, I’ll be happy
to have a way to move between windows that are simply keystrokes.