Make my “g” grep alias wonderful again
Having “g” recursive is not useful now that I use “ag” always for
recursive searches; it was preventing my using “| g” in a pipeline,
which turns out to be my most common use case of typing “g”.

There was a bit of a conflict between the “less” option “-F” and the
fact that, at least in an xterm, less prints its output over on an
alternative screen from the main one: the combination makes short output
disappear entirely and not display!  So I am adding “less -X” so that
output appears right in the main scroll buffer, and I’ll see if I like
the result.