Fix: ,simplehttpserver was serving wrong directory
Something is setting the `directory` attribute and overwriting the value
I set, so let’s switch to another name and see if it avoids collision.
Looks like it’s working so far.  Was this a change that came in with
Python 3.8?  Maybe.  Maybe I could use this later as a good example of
why never to extend a class with a subclass.