Experiment with my own New Tab page in Chrome
This avoids the expense of the built-in New Tab page, as well as the
redundancy of two search bars.  But, alas, a custom New Tab page does
not display the Bookmarks bar like the built-in New Tab, so we use a bit
of JavaScript to draw links for them anyway.