Use fzf ‘--exact’ for shell history search
Through long years of ^R searching, I expect something like ‘ls’ to only
bring up contiguous examples of ‘l’ and ‘s’; it has been distracting to
have results that merely have an ‘l’ and ‘s’ somewhere on the line.
That’s what spaces are for in an fzf search, which I can easily type.